# Conflicts:
#	modules/core/src/main/java/org/apache/ignite/configuration/NearCacheConfiguration.java

+++ modules/core/src/main/java/org/apache/ignite/internal/processors/rest/handlers/task/GridTaskResultRequest.java
@@ -163,7 +163,7 @@ public class GridTaskResultRequest implements Message {
     }
 
     /** {@inheritDoc} */
-    @Override public byte directType() {
+    @Override public short directType() {
         return 76;
     }
